% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/axisrange.R
\name{autorange_bar_y}
\alias{autorange_bar_y}
\title{autorange_bar_y}
\usage{
autorange_bar_y(fileName, Nrm = TRUE, errorbar = "SEM",
  nMajorTicks = 5, DfltZero = TRUE)
}
\arguments{
\item{fileName}{Input file name. Data should be arranged same as the input file for \code{\link{rbioplot}}.Case sensitive and be sure to type with quotation marks. Currently only takes \code{.csv} files.}

\item{Nrm}{When \code{TRUE}, normalize data to control/first group (as 1). Default is \code{TRUE}.}

\item{errorbar}{Set the type of errorbar. Options are standard error of the mean (\code{"SEM"}, \code{"standard error"}, \code{"standard error of the mean"}), or standard deviation (\code{"SD"}, \code{"standard deviation"}), case insensitive. Default is \code{"SEM"}.}

\item{nMajorTicks}{Number of major ticks intended to use for the plot. Note that the input number should be major tick number EXCLUDING 0 (or y axis lower limit if not using 0). Default is \code{5}. Note: Depending on the raw range, the last label may or may not show up due to plotting optimization, see \code{\link{rbioplot}}.}

\item{DfltZero}{When \code{TRUE}, start y axis from \code{0}. Default is \code{TRUE}.}
}
\value{
A list object containing \code{lower_limit}, \code{upper_limit}, \code{major_tick_range} and \code{minor_tick_options}.
}
\description{
A function to get custom lower/upper limit, major tick range, as well as minor tick options for y axis, based on a user-defined major tick number.
}
\examples{
\dontrun{
autorange_bar_y("data.csv", Nrm = T, errorbar = "SEM", nMajorTicks = 8, DfltZero=FALSE)
}
}
