% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sharedfunctions.R
\name{minor_tick}
\alias{minor_tick}
\title{minor_tick}
\usage{
minor_tick(major, n_minor)
}
\arguments{
\item{major}{A vector of major ticks}

\item{n_minor}{Number of minor ticks}
}
\value{
A vector containing the spaces for minor ticks
}
\description{
A function to calculate space for minor ticks
}
\examples{
\dontrun{
minor_tick(seq(1, 10, by = 2), 4)
}
}
