% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/axisrange.R
\name{nat_dvsr}
\alias{nat_dvsr}
\title{nat_dvsr}
\usage{
nat_dvsr(x)
}
\arguments{
\item{x}{An integer number}
}
\value{
A vector containing all the divisors for the input integer
}
\description{
A simple to use function to find all divisors for an integer number.
}
\examples{
\dontrun{
nat_dvsr(10)
}
}

