% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbioplot_app.R
\name{rbioplot_app}
\alias{rbioplot_app}
\title{rbioplot_app}
\usage{
rbioplot_app()
}
\description{
The web app version of \code{\link{rbioplot}}.
}
\examples{
\dontrun{
rbioplot_app() # launch the app version by running the function
}
}
