% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbioplot_curve_app.R
\name{rbioplot_curve_app}
\alias{rbioplot_curve_app}
\title{rbioplot_curve_app}
\usage{
rbioplot_curve_app()
}
\description{
The web app version of \code{\link{rbioplot_curve}}.
}
\examples{
\dontrun{
rbioplot_curve_app() # launch the app version by running the function
}
}
