% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbioplot_heatmap_app.R
\name{rbioplot_heatmap_app}
\alias{rbioplot_heatmap_app}
\title{rbioplot_heatmap_app}
\usage{
rbioplot_heatmap_app()
}
\description{
The web app version of \code{\link{rbioplot_heatmap}}.
}
\examples{
\dontrun{
rbioplot_app() # launch the app version by running the function
}
}
