% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbiostats_app.R
\name{rbiostats_app}
\alias{rbiostats_app}
\title{rbiostats_app}
\usage{
rbiostats_app()
}
\description{
The web app version of \code{\link{rbiostats}}.
}
\examples{
\dontrun{
rbiostats_app() # launch the app version by running the function
}
}
