% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbiostats.R
\name{rbiostats_normal}
\alias{rbiostats_normal}
\title{rbiostats_normal}
\usage{
rbiostats_normal(x, method = c("residual", "group"), group)
}
\arguments{
\item{x}{Input data frame or matrix for test data.}

\item{method}{Shaprio-Wilk test method. Options are \code{"residual"} and \code{"group"}. Default is \code{"residual"}.}

\item{group}{Set only when \code{method = "group"}, a factor vector for the sample data groupping.}
}
\value{
Outputs a data frame object containing normality test results to the environment.
}
\description{
A simple to use function for Shapiro-Wilk test for normality.
}
\details{
When \code{method = "residual"}, the residual for each data point is calculated using \code{residual(i) = x(i) - mean(X)} before Shapiro_Wilk test.
In such case, the test is done on all the data points regardless of groupping. Regarding \code{method = "group"}, the raw data is used. However, the test is done
according to the groupping.
}
\examples{
\dontrun{

rbiostats_normal(x = data, method = "group", group = group)

}
}
