% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frogplots.R
\name{revsort}
\alias{revsort}
\title{revsort}
\usage{
revsort(x)
}
\arguments{
\item{x}{character string with \code{"-"}}
}
\value{
Outputs a \code{string} with reversed order around \code{"-"}. Note that the function only apply to the first \code{"-"}. In the case of mulitple \code{"-"}, it will cut the rest off.
}
\description{
A function swtiches the order around the "-" symbol of a character string: from "a-b" to "b-a".
}
\examples{
\dontrun{
ab<-"a-b"
ab<-revsort(ab)
ab
}
}

