% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbioplot.R
\name{rightside_y}
\alias{rightside_y}
\title{rightside_y}
\usage{
rightside_y(ggobject)
}
\arguments{
\item{ggobject}{Input ggplot2 object.}
}
\value{
A ggplot2 object with rightside y axis
}
\description{
A function to calculate space for minor ticks
}
\examples{
\dontrun{
plt <- rightside_y(plt)
}
}
