% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbioplot_app.R
\name{set_hue}
\alias{set_hue}
\title{set_hue}
\usage{
set_hue(n)
}
\arguments{
\item{n}{Number of variables to plot.}
}
\description{
Intermediate function enables the colour change functionality for \code{\link{rbioplot_app}}.
}
\examples{
\dontrun{
col <- set_hue(n = 4) # launch the app version by running the function
}
}
