% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sharedfunctions.R
\name{shared_legend}
\alias{shared_legend}
\title{shared_legend}
\usage{
shared_legend(ggobject_list, ncol = length(ggobject_list), nrow = 1,
  position = c("bottom", "right"))
}
\arguments{
\item{ggobject_list}{Input ggplot2 object list containing multiple ggplot objectes.}

\item{ncol}{Number of columns in one figure page.}

\item{nrow}{Number of rows in one figure page.}

\item{position}{Legend position. Options are \code{"bottom"} or \code{"right"}.}
}
\value{
A grob object with mulitple ggplot2 plots in one page with shared legends.
}
\description{
A function to calculate space for minor ticks
}
\details{
This is a modified version from tidyverse: \url{https://github.com/tidyverse/ggplot2/wiki/Share-a-legend-between-two-ggplot2-grap}
}
\examples{
\dontrun{
plt <- rightside_y(plt)
}
}
