% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap.R
\name{rbioplot_heatmap}
\alias{rbioplot_heatmap}
\title{rbioplot_heatmap}
\usage{
rbioplot_heatmap(fileName, Tp = "Dunnett", Title = NULL,
  fontType = "sans", tileLow = "skyblue", tileHigh = "midnightblue",
  tileLbl = TRUE, tileLblSize = 10, tileTxtColour = "white",
  xLabel = NULL, xTickLblSize = 10, xTickItalic = FALSE, xAngle = 0,
  xAlign = 0.5, yLabel = NULL, yTickLblSize = 10, yTickItalic = FALSE,
  legendTtl = FALSE, plotWidth = 170, plotHeight = 150)
}
\arguments{
\item{fileName}{Input file name. Case sensitive and be sure to type with quotation marks. Currently only takes \code{.csv} files.}

\item{Tp}{Type of the intended statistical test. Case sensitive and be sure to type with quotation marks. Options are: "t-test", "Tukey" and "Dunnett". Default is "Dunnett".}

\item{Title}{The displayed title on top of the plot. Be sure to type with quotation marks. Default is \code{NULL}.}

\item{fontType}{The type of font in the figure. Default is "sans". For all options please refer to R font table, which is avaiable on the website: \url{http://kenstoreylab.com/?page_id=69}.}

\item{tileLow}{Set the colour for the lower limit of the heatmap. Default is \code{skyblue}. For full colour options and names, refer to the website \url{http://kenstoreylab.com/?page_id=69}.}

\item{tileHigh}{Set the colour for the upper limit of the heatmap. Default is \code{midnightblue}. For full colour options and names, refer to the website \url{http://kenstoreylab.com/?page_id=69}.}

\item{tileLbl}{Enable or disable significant notation on the tiles. Default is \code{TRUE}.}

\item{tileLblSize}{Set the font size of the tile label. Default is \code{10}.}

\item{tileTxtColour}{Set the colour of the on tile label. Default is \code{"white"}. For full colour options and names, refer to the website \url{http://kenstoreylab.com/?page_id=69}.}

\item{xLabel}{x axis label. Type with quotation marks. Default is \code{NULL}.}

\item{xTickLblSize}{Font size of x axis ticks. Default is 10.}

\item{xTickItalic}{Set x axis tick font to italic. Default is \code{FALSE}.}

\item{xAngle}{The rotation angle (degrees) of the x axis marks. Default is \code{0} - horizontal.}

\item{xAlign}{The alignment type of the x axis marks. Options are \code{0}, \code{0.5} and \code{1}. The default value at \code{0} is especially useful when \code{xAngle = 90}.}

\item{yLabel}{y axis label. Type with quotation marks. Default is \code{NULL}.}

\item{yTickLblSize}{Font size of y axis ticks. Default is 10.}

\item{yTickItalic}{Set y axis tick font to italic. Default is \code{FALSE}.}

\item{legendTtl}{Hide/Display legend title. Default is \code{FALSE}.}

\item{plotWidth}{The width of the plot (unit: mm). Default is 170. Default will fit most of the cases.}

\item{plotHeight}{The height of the plot (unit: mm). Default is 150. Default will fit most of the cases.}

\item{y_custom_tick_range}{To initiate setting the custom \code{y_upper_limit}, \code{y_lower_limit}, \code{y_major_tick_range}, \code{y_n_minor_ticks}. Default is \code{FALSE}.}

\item{y_upper_limit}{Can only be set when \code{y_custom_tick_range = TRUE}. Set custom upper limt for y axis. Value can be obtained from \code{\link{autorange_bar_y}}.}

\item{y_lower_limit}{Can only be set when \code{y_custom_tick_range = TRUE}. Set custom lower limt for y axis. Default is \code{0}. Value can be obtained from \code{\link{autorange_bar_y}}.}

\item{y_major_tick_range}{Can only be set when \code{y_custom_tick_range = TRUE}. Set custom major tick range for y axis.  Value can be obtained from \code{\link{autorange_bar_y}}.}

\item{y_n_minor_ticks}{Can only be set when \code{y_custom_tick_range = TRUE}. Set custom numbers of minor ticks. Default is \code{4}. Value can be obtained from \code{\link{autorange_bar_y}}.}
}
\value{
Outputs a \code{.csv} file with detailed metrics for the plot, including Mean, SEM and significance labels, as well as a plot image file (\code{.pdf}), with 600 dpi resolution.
}
\description{
A function for plotting simple heatmap basing on the statistical analysis of choice.
}
\examples{
\dontrun{
rbioplot("data.csv", Tp = "Tukey",
yLabel = "Relative fluorescence level")

rbioplot("data2.csv", Tp = "t-test", xAngle = -90,
xAlign=0,yLabel="Relative fluorescence level")

rbioplot("data3.csv", Tp = "Tukey",
yLabel = "Relative fluorescence level")

rbioplot("data4.csv", Tp = "Dunnett",
yLabel = "Relative fluorescence level")

rbioplot("data5.csv", Tp = "Tukey",
yLabel = "Relative fluorescence level", plotWidth = 300)

rbioplot("data8.csv", Tp = "Tukey", errorbar = "SD"
yLabel = "Relative fluorescence level",
y_custom_tick_range = TRUE, y_upper_limit = 4,
y_lower_limit = 0, y_major_tick_range = 1,
y_n_minor_ticks = 4)
}
}

